Scope create template button to any org
If you can create a template in any organization, you will see the
create template button in the workspace page.  The permission is renamed
to canCreateAnyTemplate to make it clear any org will work.

Additionally, renamed updateTemplates to updateAllTemplates to make it
clear the permission is site-wide.

I removed deleteTemplates because it is not used anywhere.